Copied the pdf-extension as suggested on Feb 21 in #3860.
This is a WIP and is not working. I don't yet understand how/when the pdf extension is loaded.

The diffences are that there are multiple mime types so I've made some assumptions on how the flow works for the mime type to be detected.

I've left the base64enciding copied from the pdf-extension, Once this is tested it needs to be moved somewhere common.

External References to attempt to understand whats going on.

https://developer.mozilla.org/en-US/docs/Web/API/Media_Source_Extensions_API
https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types
https://www.w3.org/TR/html5/embedded-content-0.html#the-audio-element
https://www.lifewire.com/audio-file-mime-types-3469485